Refactored the data access
