+++ DevOpsMetrics/DevOpsMetrics.Service/DataAccess/DeploymentFrequencyDA.cs
@@ -2,12 +2,8 @@
 using DevOpsMetrics.Service.Models.AzureDevOps;
 using DevOpsMetrics.Service.Models.Common;
 using DevOpsMetrics.Service.Models.GitHub;
-using Microsoft.Extensions.Configuration;
-using Newtonsoft.Json;
-using Newtonsoft.Json.Linq;
 using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Threading.Tasks;
 
 namespace DevOpsMetrics.Service.DataAccess
